Small adjustments to offset type

+++ EurexEti/Messages/ClipDeletionNotification.cs
@@ -93,7 +93,7 @@ namespace Eurex.EtiDerivatives.v121
 
             // --- complete header ---
 
-            BodyLen.Encode(pointer, offset, current - offset);
+            BodyLen.Encode(pointer, offset, (uint)(current - offset));
         }
 
         /// <summary>
