+++ EurexEti/Messages/DeleteAllQuoteResponse.cs
@@ -66,7 +66,7 @@ namespace Eurex.EtiDerivatives.v121
 
             // --- complete header ---
 
-            BodyLen.Encode(pointer, offset, current - offset);
+            BodyLen.Encode(pointer, offset, (uint)(current - offset));
         }
 
         /// <summary>
